requirements.txt
psycopg2-binary
requests
sqlalchemy
python-dotenv
psycopg2
fastapi
uvicorn
apscheduler
slowapi